backend/requirements.txt
# Flask==2.2.2
# tensorflow==2.13.0
# werkzeug==2.2.2
# numpy==1.23.5
# Pillow==9.3.0

tensorflow==2.18.0
Flask==3.1.0
Werkzeug==3.1.3
numpy==2.0.2
pillow==11.1.0
